+++ modules/generator/overrides_test.go
@@ -3,6 +3,7 @@ package generator
 import (
 	"time"
 
+	"github.com/grafana/tempo/pkg/sharedconfig"
 	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
@@ -14,6 +15,8 @@ type mockOverrides struct {
 	spanMetricsDimensions          []string
 	spanMetricsIntrinsicDimensions map[string]bool
 	spanMetricsFilterPolicies      []filterconfig.FilterPolicy
+	spanMetricsDimensionMappings   []sharedconfig.DimensionMappings
+	spanMetricsEnableTargetInfo    bool
 }
 
 var _ metricsGeneratorOverrides = (*mockOverrides)(nil)
@@ -56,4 +59,14 @@ func (m *mockOverrides) MetricsGeneratorProcessorSpanMetricsIntrinsicDimensions(
 
 func (m *mockOverrides) MetricsGeneratorProcessorSpanMetricsFilterPolicies(userID string) []filterconfig.FilterPolicy {
 	return m.spanMetricsFilterPolicies
+}
+
+// MetricsGeneratorProcessorSpanMetricsDimensionMappings controls custom dimension mapping
+func (m *mockOverrides) MetricsGeneratorProcessorSpanMetricsDimensionMappings(userID string) []sharedconfig.DimensionMappings {
+	return m.spanMetricsDimensionMappings
+}
+
+// MetricsGeneratorProcessorSpanMetricsEnableTargetInfo enables target_info metrics
+func (m *mockOverrides) MetricsGeneratorProcessorSpanMetricsEnableTargetInfo(userID string) bool {
+	return m.spanMetricsEnableTargetInfo
 }