+++ modules/generator/registry/test.go
@@ -23,24 +23,29 @@ func NewTestRegistry() *TestRegistry {
 	}
 }
 
-func (t *TestRegistry) NewCounter(name string, labels []string) Counter {
+func (t *TestRegistry) NewCounter(name string) Counter {
 	return &testCounter{
 		name:     name,
-		labels:   labels,
 		registry: t,
 	}
 }
 
-func (t *TestRegistry) NewLabelValues(values []string) *LabelValues {
-	return newLabelValues(values)
+func (t *TestRegistry) NewGauge(name string) Gauge {
+	return &testGauge{
+		name:     name,
+		registry: t,
+	}
 }
 
-func (t *TestRegistry) NewHistogram(name string, labels []string, buckets []float64) Histogram {
+func (t *TestRegistry) NewLabelValueCombo(labels []string, values []string) *LabelValueCombo {
+	return newLabelValueCombo(labels, values)
+}
+
+func (t *TestRegistry) NewHistogram(name string, buckets []float64) Histogram {
 	return &testHistogram{
 		nameSum:    name + "_sum",
 		nameCount:  name + "_count",
 		nameBucket: name + "_bucket",
-		labels:     labels,
 		buckets:    buckets,
 		registry:   t,
 	}
@@ -51,6 +56,13 @@ func (t *TestRegistry) addToMetric(name string, lbls labels.Labels, value float6
 		return
 	}
 	t.metrics[name+lbls.String()] += value
+}
+
+func (t *TestRegistry) setMetric(name string, lbls labels.Labels, value float64) {
+	if t == nil || t.metrics == nil {
+		return
+	}
+	t.metrics[name+lbls.String()] = value
 }
 
 // Query returns the value of the given metric. Note this is a rather naive query engine, it's only
@@ -72,41 +84,70 @@ func (t *TestRegistry) String() string {
 
 type testCounter struct {
 	name     string
-	labels   []string
 	registry *TestRegistry
 }
 
 var _ Counter = (*testCounter)(nil)
 
-func (t testCounter) Inc(values *LabelValues, value float64) {
+func (t *testCounter) Inc(labelValueCombo *LabelValueCombo, value float64) {
 	if value < 0 {
 		panic("counter can only increase")
 	}
 
-	lbls := make(labels.Labels, len(t.labels))
-	for i, label := range t.labels {
-		lbls[i] = labels.Label{Name: label, Value: values.values[i]}
+	lbls := make(labels.Labels, len(labelValueCombo.labels.names))
+	for i, label := range labelValueCombo.labels.names {
+		lbls[i] = labels.Label{Name: label, Value: labelValueCombo.labels.values[i]}
 	}
 	sort.Sort(lbls)
 
 	t.registry.addToMetric(t.name, lbls, value)
 }
 
+type testGauge struct {
+	name     string
+	registry *TestRegistry
+}
+
+var _ Gauge = (*testGauge)(nil)
+
+func (t *testGauge) Inc(labelValueCombo *LabelValueCombo, value float64) {
+	if value < 0 {
+		panic("counter can only increase")
+	}
+
+	lbls := make(labels.Labels, len(labelValueCombo.labels.names))
+	for i, label := range labelValueCombo.labels.names {
+		lbls[i] = labels.Label{Name: label, Value: labelValueCombo.labels.values[i]}
+	}
+	sort.Sort(lbls)
+
+	t.registry.addToMetric(t.name, lbls, value)
+}
+
+func (t *testGauge) Set(labelValueCombo *LabelValueCombo, value float64) {
+	lbls := make(labels.Labels, len(labelValueCombo.labels.names))
+	for i, label := range labelValueCombo.labels.names {
+		lbls[i] = labels.Label{Name: label, Value: labelValueCombo.labels.values[i]}
+	}
+	sort.Sort(lbls)
+
+	t.registry.setMetric(t.name, lbls, value)
+}
+
 type testHistogram struct {
 	nameSum    string
 	nameCount  string
 	nameBucket string
-	labels     []string
 	buckets    []float64
 	registry   *TestRegistry
 }
 
 var _ Histogram = (*testHistogram)(nil)
 
-func (t testHistogram) ObserveWithExemplar(values *LabelValues, value float64, traceID string, multiplier float64) {
-	lbls := make(labels.Labels, len(t.labels))
-	for i, label := range t.labels {
-		lbls[i] = labels.Label{Name: label, Value: values.values[i]}
+func (t *testHistogram) ObserveWithExemplar(labelValueCombo *LabelValueCombo, value float64, traceID string, multiplier float64) {
+	lbls := make(labels.Labels, len(labelValueCombo.labels.names))
+	for i, label := range labelValueCombo.labels.names {
+		lbls[i] = labels.Label{Name: label, Value: labelValueCombo.labels.values[i]}
 	}
 	sort.Sort(lbls)